Fixed bug in dialog fragment
-fixed a pretty major (and stupid) bug that I had overlooked with the
info dialog fragment.  I was originally passing an argument when
creating a new instance of the fragment and using that to determine the
title.  This is not really a good method of doing this, so I redid the
constructor to use a ‘newInstance’ method which takes the text argument
and sets that text to an arguments bundle, which then gets passed into
the onCreateView method.  Using this, and the onSaveInstanceState
methods, the fragment now handles rotation just fine.